% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_scatter.r
\name{quick_scatter}
\alias{quick_scatter}
\title{Quick scatter with identity line}
\usage{
quick_scatter(df, x, y, alpha = 0.6, add_lm = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{x, y}{Columns (unquoted)}

\item{alpha}{Point transparency}

\item{add_lm}{Add a linear fit line?}
}
\description{
Quick scatter with identity line
}
