% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_violin.r
\name{quick_violin}
\alias{quick_violin}
\title{Quick violin with median + IQR}
\usage{
quick_violin(df, x, y)
}
\arguments{
\item{df}{A data frame}

\item{x}{Grouping column (unquoted)}

\item{y}{Numeric column (unquoted)}
}
\description{
Quick violin with median + IQR
}
